Introduce ethereum-tests as a submodule
To bring this commit in (after making sure you have don't have any
unsaved changes), remove the repository from `tests/ethereum-tests/`.
Rebase onto or checkout this branch, and then run `git submodule
update --init` to clone the repository, and to setup `.git/config`.

This in nicer than simply cloning the git repository to that location
because:

 * It explicilty documents the dependency -- no need to go hunting
   around the internet to find the correct repositorty.
 * It documents the exact version of the tests we're running against.
   Since git-submodules stores the SHA of the commit to use,
   we're guaranteed that all developers are running the same version.
   We can then explicitly upgrade to a newer version as needed.