simplifying testing harness
There were several things which were "duplication" of code in the testing harness.
Now that we have a solid test-set setup going, it seemed time to clean it up.

First, addresses are converted to their Int representation, which makes it easier to understand the diff output of tests.
Second, independent of whether the command is `check` or `load`, the sub-parts of the command are handled uniformly.