data.md: Add memo attribute to symbolic ByteArray functions
The Haskell backend will use the memo attribute to recognize which functions
should be memoized. The attribute is added to symbolic ByteArray functions which
are linear in size, such as `#sizeByteArray`, which tend to be called on every
program step.

This should have no effect on the Java backend, which I understand memoizes
everything. This should have no effect on the LLVM backend because only the
symbolic definitions are modified.