Less restrictive rules for Foundry cheatcodes.
Previously, cheatcodes where matched against the following CALL instruction:

  CALL _ CHEAT_ADDR 0 ARGSTART _ARGWIDTH 0 0

In this commit, this is changed as follows:

  CALL _ CHEAT_ADDR _VALUE ARGSTART _ARGWIDTH _RETSTART _RETWIDTH

In particular:
1. The memory region that contains the return value may now be arbitrary
  (_RETSTART and _RETWIDTH)
2. An arbitrary amount of ether may be sent when calling the cheatcode (_VALUE)

Because RETSTART and RETWIDTH were initially required to be zero, the K rules
where actually never applied. This is because the bytecode emitted by the
Solidity compiler always sets RETSTART to a non-zero value, even for functions
that do not return any value.

Allowing arbitrary ether to be sent along with cheatcodes ensures we are
consistent with Foundry. Note that the ether is not actually transferred but
simply ignored. This, too, is consistent with Foundry.