Isolated and fixed some proofs for arithemtic functions used in Vat.
RPC call returned `k` cell and `wordstack` cell that was past the target `k` cell and `wordstack` cell. Therefore the prover target state could not be matched. Editting these so they matched the returned state at the branch from the RPC call, and ammending the gas costs allows a match.