Isolated and fixed some proofs for arithemtic functions used in Vat. (#1695)
* Isolated and fixed some proofs for arithemtic functions used in Vat.

RPC call returned `k` cell and `wordstack` cell that was past the target `k` cell and `wordstack` cell. Therefore the prover target state could not be matched. Editting these so they matched the returned state at the branch from the RPC call, and ammending the gas costs allows a match.

* Set Version: 1.0.74

* flake.{nix,lock}: update Nix derivations

* Set Version: 1.0.78

* kevm-pyk/: sync poetry files

* flake.{nix,lock}: update Nix derivations

* Set Version: 1.0.79

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Everett Hildenbrandt <everett.hildenbrandt@gmail.com>