Constructor implementation of infinite gas.
`Gas` sort is added to `evm.md` as a supersort of `Int`, this allows all rules in `evm.md` to remain unchanged as `Gas` can be inferred as an `Int` from the operations.
`infinite-gas.md` extends `Gas` with infinite gas `#gas(Int)`. `#deductgas` and `#execGas` needed additional constructors as `#gas(_)` would not longer match on int.
It appears with the additional rules that the arithmetic operations are not necessary to pass all the proofs in `smoke` except `transfer-success-2-spec.k` - I am unsure why this is not passing, the stopped state appears equivalent to the target state but with `(looped back)`. Adding this PR for comment on the changes.