Added `Gas` as sort of `WordStack`.
This allows proofs that were blocking on `GAS` instruction to go through, as `Gas` type from `<gas>` cell  needs to be pushed to the top of the stack.
Subtraction mcd proofs are currently failing and I cannot track the reason why.

Please enter the commit message for your changes. Lines starting