Current constructor implementation of infinite gas fails this test.
Explanation:
`gasleft()` compiles to `GAS` opcode which adds the contents of `<gas>` as an `Int` to the stack. A `#gas(G)` is unwrapped to a `G`. This loses the information that these numbers came from infinite gas, and a comparison of the two does not consider them as `#gas(_)` and so will always fail.