Refactor blockchain-k-plugin build (#1865)
As we are starting to test the booster with more and more kevm proofs, we are hitting new corner cases with the LLVM dynamic library, which the booster uses for concrete execution. Specifically, we have come across the following error:

```
/tmp/hs-booster-proxy-bfeaddc9bbbd2ef1/.log: withFile: user error (dlopen: out/kompiled-llvm/interpreter.so: undefined symbol: _ZTVN8CryptoPP6SHA512E)
```

It turns out the linker is being too aggressive when pruning seemingly unused symbols from the dynamic library which results in the above error. The fix is to compile the offending library statically, producing a`.a` file and then instead of passing e.g. `-lcryptopp -L .../libcryptopp/lib`, we instead provide the static `.../libcryptopp/lib/libcryptopp.a` via `-ccopts` in `kevm-pyk`.

Since we now rely on static libraries for `libcryptopp`, `libff` and `libsecp256k1`, we build them from source, insterad of relying on the nix/system provided shared libraries.