Add booster tests to CI (#1890)
The booster backend is now getting to the state where we can start testing against it for foundry proofs. This PR adds a CI run for the foundry tests with booster. Running with booster can be enabled for kevm foundry by first compiling with `kevm foundry-kompile ... --with-llvm-library` and then `kevm foundry-prove ... --use-booster`