test lemmas for isStaticType and resulting encodeArgs values
In PR 2371, an `[owise]` attribute was added to `#isStaticType`.

The original reported issue was
[hs-backend-booster/issues/558](https://github.com/runtimeverification/hs-backend-booster/issues/558),
where a wrong rule was applied because of the missing `[owise]` (the
LLVM backend library evaluated the side condition to the wrong value).

This PR adds two lemmas to test the behaviour. On KEVM versions before
PR 2371 the tests fail when using `--use-booster` but succeed without it.