Change how json is parsed in pytest harness (#5)
Previously we were parsing the json tests in python, then converting
them to KORE and passing them to the backend which then used the kore
parser to parse them. This is fairly inefficient, because we end up
parsing things twice.

The solution is pretty straightforward. We simply send the entire json
file to the backend as a single `String`, then invoke the llvm backend's
json parser, which is extremely fast, on the resulting term. It's a
fairly basic change, but it significantly improves performance.

Time on the base branch:
```
real    6m21.214s
user    75m41.967s
sys     3m35.565s
```
Time on the branch containing the change:
```
real    5m23.010s
user    46m6.622s
sys     2m47.065s
```