First draft of the USL API (#8)
This is a first draft for the low-level API that will be used to
communicate with the DSL.

Note that many of the data types used in this draft are simply 64-bit
integers. These are placeholders and I welcome your input as to what
would be a recommended data type for each such placeholder.