install-build-deps: Automate installation of required system packages
These used to be listed in README.md; some notes on the various
systems still remain there.

The framework for this script is taken from the same script in [k].

This has been tested on Ubuntu 22.04 and recent Arch. The MacOS code from
the [k] version has been tested on MacOS, and this is so similar (including
the use of `stack`) that there should be no issues with it. The Arch
support is untested, but is so simple it's hard to see what could go wrong.

[k]: https://github.com/runtimeverification/k