Introducing summary for `#loadProgram` to avoid computing `#computeValidJumpDests(PGM)` (#20)
Fixes https://github.com/Pi-Squared-Inc/pi2/issues/2119

This PR contains the summaries for all contracts in
https://github.com/Pi-Squared-Inc/pi2/tree/main/generation/evm-benchmarks
except for
[add11.json](https://github.com/Pi-Squared-Inc/pi2/blob/main/generation/evm-benchmarks/add11.json)
and
[sumTo10.evm](https://github.com/Pi-Squared-Inc/pi2/blob/main/generation/evm-benchmarks/sumTo10.evm).
The first contains more than one call for `#computeValidJumpDests`, and
the second I'll need more info to execute it to write the summary.
@alexoltean61, please let me know if they are important to you, and I
can double-check if we can summarize them as well.