Fix `#codeDeposit` rule to set the output correctly when status is different from success (#25)
This PR fixes a behavior where we didn't persist the output from the
execution of the `revert` opcode. Without this issue, tests that rely on
the result of this operation were failing with gas mismatch, i.e.,
operations that were supposed to happen didn't execute. This PR
addresses the issue and doesn't add any regression to the test suite.