Add <worldState> cell for capturing side-effects of impure hooks
During concrete execution these hooks execute to `.K`. This does not work during
symbolic execution, as the side-effects of the hooks would not be accounted for.

In the VLM repo, we have modified the VLM hooks for symbolic execution to instead be
uninterpreted functions that return the WorldState sort.
Here, we create a <worldState> cell to store the side-effects, while unblocking
execution in the <k> cell.