KZG* precompiles: Use ==Bytes instead of ==K
We introduce the ==Bytes construct, because K seems to have difficulty
figuring out that a term of sort K passed reached via executing the semantics
is the same as that introduced via pyk. The reason for this seems related
to the first being parsed as a KSequence, i.e. giveing us `X:Bytes ~> .K`.