Update submodule for blockchain-k-plugin (#59)
This PR contains the change to the version of the blockchain-k-plugin
used by the evm semantics. This version of the plugin is missing some of
the symbols needed to link the crypto hooks because it is expected that
the semantics will now be built into a static library and thus be able
to make use of the versions of the same symbols provided by the rust
libraries that use the same code for their own cryptographic
implementations.

This version of the semantics will probably not build with the latest
default branch of revm, but it is intended to make the revm PR
https://github.com/Pi-Squared-Inc/revm/pull/46 function correctly.

---------

Co-authored-by: Roberto Rosmaninho <roberto.rosmaninho@pi2.network>