`EVMC_PRECOMPILE_OOG` error type (#69)
This PR adds the `EVMC_PRECOMPILE_OOG` error type, which is needed for
op-revm, and returns is when appropriate (out of gas when executing a
precompiled contract with an optimism schedule).

To return the correct error, `EVMC_PRECOMPILE_OOG` or `EVMC_Out_Of_Gas`,
we need to know whether a particular OOG is during execution of a
precompiled contract - to that end, we added a cell that contains this
information.

The remaining 2 op-revm unit tests that are failing will be investigated
as part of new PRs.