Allow for random/non consecutive precompile addresses. (#72)
This PR addresses issue Pi-Squared-Inc/pi2/issues/2837

- The proposed `syntax OpCode ::= getPrecompile(MInt)` already exists,
it is this one:

https://github.com/Pi-Squared-Inc/evm-semantics/blob/f4bf49bb7415599fbf2073266d95b6c9b709c72e/kevm-pyk/src/kevm_pyk/kproj/evm-semantics/evm.md?plain=1#L1244
We will extend it as new precompiles are added, e.g., 
```
    rule #precompiled(256p256) => P256VERIFY
```
- The proposed `syntax Bool ::= isPrecompile(MInt, Schedule)`, that
should check whether a MInt is an address of a precompiled contract in a
Schedule corresponds to the existing `syntax Bool ::=
#isPrecompiledAccount ( MInt{256} , Schedule )`. I replaced the
implementation according to our discussion: It is defined for all pairs
of `MInt`, `Schedule` for which it is true, and false otherwise. For
example, we will include the following rules
```
rule #isPrecompiledAccount(256p256, FJORD) => true
rule #isPrecompiledAccount(256p256, GRANITE) => true
rule #isPrecompiledAccount(256p256, HOLOCENE) => true
rule #isPrecompiledAccount(256p256, ISTHMUS) => true
```
to enable `P256VERIFY` on schedules `FJORD` onwards.

We also remove `#precompiledAccountsSet`, since
- I found no uses of it
- it used `#precompiledAccountsUB` which was no longer accurate and also
deleted
If we need it, let me know and I will redefine it, using a rule per
schedule.